Datasets about mental disorder prevalence by country
The CSV file contains data about mental disorders for different countries over multiple years. The first attribute is about Disability Adjusted Life Years (DALYs) which measures the mortality associated with a disorder - mental disorders in this case. The second and third attributes measure the prevalence of Eating Disorders and Depressive disorders as a percentage.